requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
CurrencyConverter==0.18.5
Django==5.2
django-celery-results==2.5.1
django-environ==0.12.0
django-redis==5.4.0
djangorestframework==3.16.0
idna==3.10
kombu==5.5.2
prompt_toolkit==3.0.50
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
